+++ controler/RoleControler.php
@@ -38,9 +38,10 @@ class RoleControler extends PastellControler
     public function detailAction(): void
     {
         $this->checkDroitFor(EntiteSQL::ID_E_ENTITE_RACINE, DroitService::DROIT_ROLE, DroitType::LECTURE);
-        $this->setViewParameter('role', $this->getGetInfo()->get('role'));
+        $role_id = $this->getGetInfo()->get('role');
+        $this->setViewParameter('role', $role_id);
         $this->setDroitViewParameter(EntiteSQL::ID_E_ENTITE_RACINE, DroitService::DROIT_ROLE, DroitType::EDITION);
-        $this->setViewParameter('role_info', $this->getRoleSQL()->getInfo($this->getViewParameterOrObject('role')));
+        $this->setViewParameter('role_info', $this->getRoleSQL()->getInfo($role_id));
 
         $roleDroit = $this->getInstance(RoleDroit::class);